EEPROM structure changes to allow more switches (9bits instead of 8bits)

+++ radio/src/opentx.h
@@ -423,7 +423,7 @@
   typedef int32_t rotenc_t;
   typedef int32_t getvalue_t;
   typedef uint32_t mixsrc_t;
-  typedef int8_t swsrc_t;
+  typedef int32_t swsrc_t;
 #else
   #define tmr10ms_t uint16_t
   extern volatile tmr10ms_t g_tmr10ms;
